Revert "Move funtion imports"
This reverts commit cb8124e5
