+++ pandas/core/array_algos/masked_reductions.py
@@ -8,7 +8,7 @@ from typing import Callable
 import numpy as np
 
 from pandas._libs import missing as libmissing
-from pandas.compat import np_version_under1p17
+from pandas.compat.numpy import np_version_under1p17
 
 from pandas.core.nanops import check_below_min_count
